F formatting code

+++ HYElevation/ElevationFetcher.m
@@ -15,7 +15,7 @@
     NSString *path = [[NSBundle mainBundle] pathForResource:@"elevationData" ofType:@"archive"];
     NSArray *testData=[NSKeyedUnarchiver unarchiveObjectWithFile:path];
     HYElevationPoint *temp;
-    for (int i=0; i < testData.count; i++) {
+    for (NSUInteger i=0; i < testData.count; i++) {
         temp = (HYElevationPoint *)testData[i];
         if (i == 60) temp.elevationPointName = @"aPoint";
         else if (i == 120) temp.elevationPointName = @"otherPoint";
